Extract plain vector index
